+++ sdk/eventgrid/Azure.Messaging.EventGrid/tests/EventGridClientLiveTests.cs
@@ -33,6 +33,27 @@ namespace Azure.Messaging.EventGrid.Tests
         }
 
         [RecordedTest]
+        public async Task CanPublishSingleEvent()
+        {
+            EventGridPublisherClientOptions options = InstrumentClientOptions(new EventGridPublisherClientOptions());
+            EventGridPublisherClient client = InstrumentClient(
+                new EventGridPublisherClient(
+                    new Uri(TestEnvironment.TopicHost),
+                    new AzureKeyCredential(TestEnvironment.TopicKey),
+                    options));
+            await client.SendEventAsync(
+                new EventGridEvent(
+                        "Subject",
+                        "Microsoft.MockPublisher.TestEvent",
+                        "1.0",
+                        "hello")
+                {
+                    Id = Recording.Random.NewGuid().ToString(),
+                    EventTime = Recording.Now
+                });
+        }
+
+        [RecordedTest]
         public async Task CanPublishEventWithCustomObjectPayload()
         {
             EventGridPublisherClientOptions options = InstrumentClientOptions(new EventGridPublisherClientOptions());
@@ -152,6 +173,28 @@ namespace Azure.Messaging.EventGrid.Tests
             }
 
             await client.SendEventsAsync(eventsList);
+        }
+
+        [RecordedTest]
+        public async Task CanPublishSingleCloudEvent()
+        {
+            EventGridPublisherClientOptions options = InstrumentClientOptions(new EventGridPublisherClientOptions());
+            EventGridPublisherClient client = InstrumentClient(
+                new EventGridPublisherClient(
+                    new Uri(TestEnvironment.CloudEventTopicHost),
+                    new AzureKeyCredential(TestEnvironment.CloudEventTopicKey),
+                    options));
+
+            await client.SendEventAsync(
+                new CloudEvent(
+                        "record",
+                        "Microsoft.MockPublisher.TestEvent",
+                        null)
+            {
+                Id = Recording.Random.NewGuid().ToString(),
+                Subject = "Subject",
+                Time = Recording.Now
+            });
         }
 
         [RecordedTest]
@@ -365,6 +408,27 @@ namespace Azure.Messaging.EventGrid.Tests
                     new AzureKeyCredential(TestEnvironment.CustomEventTopicKey),
                     options));
             await client.SendEventsAsync(GetCustomEventsList());
+        }
+
+        [RecordedTest]
+        public async Task CanPublishSingleCustomEvent()
+        {
+            EventGridPublisherClientOptions options = InstrumentClientOptions(new EventGridPublisherClientOptions());
+            EventGridPublisherClient client = InstrumentClient(
+                new EventGridPublisherClient(
+                    new Uri(TestEnvironment.CustomEventTopicHost),
+                    new AzureKeyCredential(TestEnvironment.CustomEventTopicKey),
+                    options));
+            await client.SendEventAsync(
+                new BinaryData(new TestEvent()
+            {
+                DataVersion = "1.0",
+                EventTime = Recording.Now,
+                EventType = "Microsoft.MockPublisher.TestEvent",
+                Id = Recording.Random.NewGuid().ToString(),
+                Subject = "Subject",
+                Topic = "Topic"
+            }));
         }
 
         [RecordedTest]